Enable including crop marks
I'm printing a table because then I can see what the cards look like,
but when I print this on paper and cut it, if the cut is off by a tent
of a millimeter you're going to see an ugly black line at the edge of
the card. I don't need the table for cutting, I need crop marks. So
enable adding crop marks. But for visual inspection on my computer, I
do want the table. So allow toggling both.